Now with proper UV support !
Only one UV map though... And only one material.
Still, that should do the trick for simple items.

This also change nothing in the current format,
meaning that I'll just have to add an URL input
for the main texture, in the loader, to support
this.

Also, I didn't check about the version number
I put in the addon... I'll have to bump the release
version to 1.0 - 1.1...

Well, it's not like the version 1.0 didn't work
though, so... yeah, we'll start from 1.0 and go
towards 1.1 now !

So, next release : 1.1 !

Signed-off-by: Voyage <voyage@miouyouyou.fr>